+++ pkg/options/options.go
@@ -102,55 +102,6 @@ type Js struct {
 	Overwrite    bool        `help:"Whether to ignore and overwrite existing uploads with same identifier, rather than failing if a matching file exists"`
 }
 
-type ReactNative struct {
-	Path            utils.Paths                `arg:"" name:"path" help:"The path to the root of the React Native project to upload files from" type:"path" default:"."`
-	ProjectRoot     string                     `help:"The path to strip from the beginning of source file names referenced in stacktraces on the BugSnag dashboard" type:"path"`
-	Shared          ReactNativeShared          `embed:""`
-	AndroidSpecific ReactNativeAndroidSpecific `embed:"" prefix:"android-"`
-	IosSpecific     ReactNativeIosSpecific     `embed:"" prefix:"ios-"`
-	Overwrite       bool                       `help:"Whether to ignore and overwrite existing uploads with same identifier, rather than failing if a matching file exists"`
-}
-
-type ReactNativeAndroidSpecific struct {
-	AppManifest string `help:"The path to a manifest file (AndroidManifest.xml) from which to obtain build information" type:"path"`
-	Variant     string `help:"The build type/flavor (e.g. debug, release) used to disambiguate the between built files when searching the project directory"`
-	VersionCode string `help:"The version code of this build of the application"`
-}
-
-type ReactNativeAndroid struct {
-	Path        utils.Paths `arg:"" name:"path" help:"The path to the root of the React Native project to upload files from" type:"path" default:"."`
-	ProjectRoot string      `help:"The path to strip from the beginning of source file names referenced in stacktraces on the BugSnag dashboard" type:"path"`
-
-	ReactNative ReactNativeShared          `embed:""`
-	Android     ReactNativeAndroidSpecific `embed:""`
-	Overwrite   bool                       `help:"Whether to ignore and overwrite existing uploads with same identifier, rather than failing if a matching file exists"`
-}
-
-type ReactNativeShared struct {
-	Bundle       string `help:"The path to the bundled JavaScript file to upload" type:"path"`
-	CodeBundleId string `help:"A unique identifier for the JavaScript bundle"`
-	Dev          bool   `help:"Indicates whether this is a debug or release build"`
-	SourceMap    string `help:"The path to the source map file to upload" type:"path"`
-	VersionName  string `help:"The version of the application"`
-}
-
-type ReactNativeIosSpecific struct {
-	BundleVersion string     `help:"The bundle version of this build of the application (Apple platforms only)"`
-	Plist         string     `help:"The path to a .plist file from which to obtain build information" type:"path"`
-	Scheme        string     `help:"The name of the Xcode options.Ios.Scheme used to build the application"`
-	XcodeProject  string     `help:"The path to an Xcode project, workspace or containing directory from which to obtain build information" type:"path"`
-	XcarchivePath utils.Path `help:"The path to the Xcode archive to process if it has been exported" type:"path"`
-}
-
-type ReactNativeIos struct {
-	Path        utils.Paths `arg:"" name:"path" help:"The path to the root of the React Native project to upload files from" type:"path" default:"."`
-	ProjectRoot string      `help:"The path to strip from the beginning of source file names referenced in stacktraces on the BugSnag dashboard" type:"path"`
-
-	ReactNative ReactNativeShared      `embed:""`
-	Ios         ReactNativeIosSpecific `embed:""`
-	Overwrite   bool                   `help:"Whether to ignore and overwrite existing uploads with same identifier, rather than failing if a matching file exists"`
-}
-
 type Breakpad struct {
 	Path            utils.Paths `arg:"" name:"path" help:"The path to the symbol files (.sym) to upload (or directory containing them)" type:"path"`
 	CpuArch         string      `help:"The CPU architecture that the module was built for"`
@@ -171,22 +122,23 @@ type Upload struct {
 	UploadAPIRootUrl string `help:"The upload server hostname, optionally containing port number"`
 
 	// required options
-	All                DiscoverAndUploadAny   `cmd:"" help:"Upload any symbol/mapping files"`
-	AndroidAab         AndroidAabMapping      `cmd:"" help:"Process and upload application bundle files for Android"`
-	AndroidNdk         AndroidNdkMapping      `cmd:"" help:"Process and upload NDK symbol files for Android"`
-	AndroidProguard    AndroidProguardMapping `cmd:"" help:"Process and upload Proguard/R8 mapping files for Android"`
-	DartSymbol         DartSymbol             `cmd:"" help:"Process and upload symbol files for Flutter" name:"dart"`
-	XcodeBuild         XcodeBuild             `cmd:"" help:"Upload dSYMs for iOS from a build"`
-	Dsym               Dsym                   `cmd:"" help:"(deprecated) Upload dSYMs for iOS"`
-	XcodeArchive       XcodeArchive           `cmd:"" help:"Upload dSYMs for iOS from a Xcode archive"`
-	Js                 Js                     `cmd:"" help:"Upload source maps for JavaScript"`
-	ReactNative        ReactNative            `cmd:"" help:"Upload source maps for React Native"`
-	ReactNativeAndroid ReactNativeAndroid     `cmd:"" help:"Upload source maps for React Native Android"`
-	ReactNativeIos     ReactNativeIos         `cmd:"" help:"Upload source maps for React Native iOS"`
-	UnityAndroid       UnityAndroid           `cmd:"" help:"Upload Android mappings and NDK symbol files from Unity projects"`
-	UnityIos           UnityIos               `cmd:"" help:"Upload iOS mappings and dSYMs from Unity projects"`
-	Breakpad           Breakpad               `cmd:"" help:"Upload breakpad .sym files"`
-	Linux              LinuxOptions           `cmd:"" help:"Upload symbol/mapping files"`
+	All                   DiscoverAndUploadAny   `cmd:"" help:"Upload any symbol/mapping files"`
+	AndroidAab            AndroidAabMapping      `cmd:"" help:"Process and upload application bundle files for Android"`
+	AndroidNdk            AndroidNdkMapping      `cmd:"" help:"Process and upload NDK symbol files for Android"`
+	AndroidProguard       AndroidProguardMapping `cmd:"" help:"Process and upload Proguard/R8 mapping files for Android"`
+	DartSymbol            DartSymbol             `cmd:"" help:"Process and upload symbol files for Flutter" name:"dart"`
+	XcodeBuild            XcodeBuild             `cmd:"" help:"Upload dSYMs for iOS from a build"`
+	Dsym                  Dsym                   `cmd:"" help:"(deprecated) Upload dSYMs for iOS"`
+	XcodeArchive          XcodeArchive           `cmd:"" help:"Upload dSYMs for iOS from a Xcode archive"`
+	Js                    Js                     `cmd:"" help:"Upload source maps for JavaScript"`
+	ReactNative           ReactNative            `cmd:"" help:"Upload source maps for React Native"`
+	ReactNativeAndroid    ReactNativeAndroid     `cmd:"" help:"Upload source maps for React Native Android"`
+	ReactNativeIos        ReactNativeIos         `cmd:"" help:"Upload source maps for React Native iOS"`
+	ReactNativeSourcemaps ReactNativeSourcemaps  `cmd:"" help:"Upload source maps for React Native"`
+	UnityAndroid          UnityAndroid           `cmd:"" help:"Upload Android mappings and NDK symbol files from Unity projects"`
+	UnityIos              UnityIos               `cmd:"" help:"Upload iOS mappings and dSYMs from Unity projects"`
+	Breakpad              Breakpad               `cmd:"" help:"Upload breakpad .sym files"`
+	Linux                 LinuxOptions           `cmd:"" help:"Upload symbol/mapping files"`
 }
 
 // Unique CLI options