+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## Unreleased
 
+### Added
+
+- Add support for uploading React Native Sourcemaps outside of Android or iOS builds [#241](https://github.com/bugsnag/bugsnag-cli/pull/241)
+
 ### Changed
 
 - `bugsnag-cli-placeholder` now throws an `Error` instead of just logging to the console. This immediately stops execution and should return a stack trace if the CLI is not successfully installed [#239](https://github.com/bugsnag/bugsnag-cli/pull/239)